django_backend/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
Django==1.9
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.3.3
mongoengine==0.15.0
pymongo==3.6.0
pytz==2017.3
idna==2.6
oauthlib==2.0.6
python3-openid==3.1.0
requests==2.18.4
requests-oauthlib==0.8.0
urllib3==1.22
